+++ src/main/java/Command.java
@@ -1,15 +1,3 @@
-enum CommandType {
-    ADD_TODO,
-    ADD_DEADLINE,
-    ADD_EVENT,
-    DELETE,
-    MARK,
-    UNMARK,
-    LIST,
-    BYE,
-    INVALID
-}
-
 public class Command {
     final CommandType type;
     String maybeArgument = null;
