+++ src/main/java/CommandType.java
@@ -1,0 +1,11 @@
+public enum CommandType {
+    ADD_TODO,
+    ADD_DEADLINE,
+    ADD_EVENT,
+    DELETE,
+    MARK,
+    UNMARK,
+    LIST,
+    BYE,
+    INVALID
+}
