+++ src/main/java/task/TaskException.java
@@ -1,0 +1,8 @@
+package task;
+
+public class TaskException extends Exception {
+    final TaskExceptionType exceptionType;
+    TaskException(TaskExceptionType exceptionType) {
+        this.exceptionType = exceptionType;
+    }
+}
