+++ src/main/java/task/TaskExceptionType.java
@@ -1,0 +1,7 @@
+package task;
+
+public enum TaskExceptionType {
+    NO_ARGUMENTS_PROVIDED,
+    ARGUMENTS_MISSING,
+    UNRECOGNIZED_ARGUMENT,
+}
